requirements.txt
accelerate>=0.30.0
numpy>=1.24
pillow>=10.0.0
sentence-transformers>=3.0.0
transformers>=4.45.0
sentencepiece>=0.2.0
protobuf>=5.0.0
datasets>=2.20.0
